+++ app/Tables/Builders/UserTable.php
@@ -12,9 +12,11 @@ class UserTable extends Table
     public function query()
     {
         return User::select(\DB::raw(
-                'users.id, users.id as "dtRowId", owners.name as owner, users.first_name, users.last_name, users.phone,
-                users.email, roles.name as role, users.is_active'
+                'users.id, users.id as "dtRowId", avatars.id as avatarId, owners.name as owner,
+                users.first_name, users.last_name, users.phone, users.email, roles.name as role,
+                users.is_active'
             ))->join('owners', 'users.owner_id', '=', 'owners.id')
-            ->join('roles', 'users.role_id', '=', 'roles.id');
+            ->join('roles', 'users.role_id', '=', 'roles.id')
+            ->leftJoin('avatars', 'users.id', '=', 'avatars.user_id');
     }
 }