+++ config/enso/addresses.php
@@ -19,10 +19,31 @@ return [
         'Bend' => 'Bend',
         'Square' => 'Square',
     ],
+    'buildingTypes' => [
+        'Offices' => 'Offices',
+        'Residential' => 'Residential',
+        'Comercial' => 'Comercial',
+        'Industrial' => 'Industrial',
+    ],
     'validations' => [
         'street' => 'required',
         'city' => 'required',
         'country_id' => 'required',
+    ],
+    'resource' => [
+        'administrativeArea' => 'administrative_area',
+        'city' => 'city',
+        'subAdministrativeArea' => 'sub_administrative_area',
+        'streetType' => 'street_type',
+        'street' => 'street',
+        'building' => 'building',
+        'entry' => 'entry',
+        'floor' => 'floor',
+        'apartment' => 'apartment',
+        'buildingType' => 'building_type',
+        'postalArea' => 'postal_area',
+        'obs' => 'obs',
+        'isDefault' => 'is_default',
     ],
     'loggableMorph' => [
         'addressable' => [