+++ config/enso/trackWho.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+    'resource' => ['name' => 'fullName'],
+];