+++ database/seeds/UserSeeder.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\User;
 use App\Owner;
 use Illuminate\Database\Seeder;
 use LaravelEnso\RoleManager\app\Models\Role;
@@ -8,7 +9,7 @@ class UserSeeder extends Seeder
 {
     public function run()
     {
-        \DB::table('users')->insert([
+        $id = \DB::table('users')->insertGetId([
             'email' => 'admin@laravel-enso.com',
             'first_name' => 'Admin',
             'password' => '$2y$10$06TrEefmqWBO7xghm2PUzeF/O0wcawFUv8TKYq.NF6Dsa0Pnmd/F2',
@@ -16,5 +17,8 @@ class UserSeeder extends Seeder
             'owner_id' => Owner::whereName('Admin')->first()->id,
             'is_active' => true,
         ]);
+
+        auth()->onceUsingId($id);
+        User::find($id)->generateAvatar();
     }
 }